do conditional check befor installing lnd or btc

+++ shell_scripts/install_lnd.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 
-    
+if systemctl is-active lnd; then
+  echo "LND is running"
+  exit 1
+
+echo "starting to install lnd"
 mkdir -p ~/birkeland/code
 cd ~/birkeland/code && rm -rf lnd_code
 mkdir -p lnd_code && cd lnd_code
